+++ code/Samples/Input/Sensors/Sensors.cc
@@ -55,8 +55,8 @@ SensorsApp::OnInit() {
     this->drawState = Gfx::CreateResource(dss);
 
     // setup transform matrices
-    const float32 fbWidth = Gfx::DisplayAttrs().FramebufferWidth;
-    const float32 fbHeight = Gfx::DisplayAttrs().FramebufferHeight;
+    const float32 fbWidth = (const float32) Gfx::DisplayAttrs().FramebufferWidth;
+    const float32 fbHeight = (const float32) Gfx::DisplayAttrs().FramebufferHeight;
     this->proj = glm::perspectiveFov(glm::radians(45.0f), fbWidth, fbHeight, 0.1f, 100.0f);
     return App::OnInit();
 }
